fix: Rename and move to lib
