+++ lib/database.ts
@@ -1,4 +1,4 @@
-import { prisma } from "./prisma";
+import { prisma } from "./prismaSetup";
 
 export async function createMetadata(
   senderMail: string,
